% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_sample.R
\name{getPortmanteaus}
\alias{getPortmanteaus}
\title{Gets portmanteaus for a given word (all known information).}
\usage{
getPortmanteaus(word = NULL, lang = NULL, maxResults = NULL)
}
\arguments{
\item{word}{String. A word to retrieve rhymes for - e.g. 'mellow'}

\item{lang}{String (optional). The language the word is written in ('en, 'de', 'es', 'fr', 'ru')}

\item{maxResults}{Integer (optional). The maximum amount of results to retrieve (if left null, Rhymebrain will impose it's own limit depending on how many "good sounding rhymes" exist for the word)}
}
\value{
Dataframe of word information, source word, alternative spelling. More information can be found here: <https://rhymebrain.com/api.html#portmanteau>
}
\description{
Gets portmanteaus for a given word (all known information).
}
\examples{
\dontrun{
getPortmanteaus('mellow','en',10)
}
}
