% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_sample.R
\name{getRhymes}
\alias{getRhymes}
\title{Gets rhyming words for a given word.}
\usage{
getRhymes(word = NULL, lang = NULL, maxResults = NULL)
}
\arguments{
\item{word}{String. A word to retrieve rhymes for - e.g. 'mellow'}

\item{lang}{String (optional). The language the word is written in ('en', 'de', 'es', 'fr', 'ru')}

\item{maxResults}{Integer (optional). The maximum amount of results to retrieve (if left null, Rhymebrain will impose it's own limit depending on how many "good sounding rhymes" exist for the word)}
}
\value{
Dataframe of rhyming words, their RhymeRank score (300 = perfect), flags, syllables and word commonality frequency. More information can be found at the API site: <https://rhymebrain.com/api.html#rhyme>
}
\description{
Gets rhyming words for a given word.
}
\examples{
\dontrun{
getRhymes('mellow','en',10)
}
}
