% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_sample.R
\name{getWordInfo}
\alias{getWordInfo}
\title{Gets word info for a given word.}
\usage{
getWordInfo(word = NULL, lang = NULL)
}
\arguments{
\item{word}{String. A word to retrieve rhymes for - e.g. 'mellow'}

\item{lang}{String (optional). The language the word is written in ('en', 'de', 'es', 'fr', 'ru')}
}
\value{
List of Information for the given word, given an optional language parameter. More information found here: <https://rhymebrain.com/api.html#pron>
}
\description{
Gets word info for a given word.
}
\examples{
\dontrun{
getWordInfo('mellow','en')
}
}
