% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_sample.R
\name{search_ban}
\alias{search_ban}
\alias{reverse_search_ban}
\title{Search the BAN}
\usage{
search_ban(
  q = NULL,
  limit = NULL,
  autocomplete = NULL,
  lat = NULL,
  lon = NULL,
  type = NULL,
  postcode = NULL,
  citycode = NULL
)

reverse_search_ban(lat = NULL, lon = NULL)
}
\arguments{
\item{q}{text search}

\item{limit}{maximum number of results to return}

\item{autocomplete}{autocompletion}

\item{lat}{latitude}

\item{lon}{longitude}

\item{type}{type of elements to return (housenumber, street,place, municipality)}

\item{postcode}{Postal code}

\item{citycode}{INSEE code}
}
\value{
the results from the search
}
\description{
Search the BAN
}
\examples{
\dontrun{
search_ban("Rennes")
reverse_search_ban("48.11", "-1.68")
}
}
